+++ sandbox/08-hooks/04-effect/src/App.tsx
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import TimerPage from './TimerPage';
+import { Button, Card, Icon, Statistic } from 'semantic-ui-react';
 
 import './App.css';
 
@@ -17,41 +17,50 @@ class App extends Component<{}, AppState> {
     this.state = { timeLeft: LIMIT };
   }
 
-  // componentDidMount で setInterval
-  componentDidMount() {
-    this.timerId = setInterval(this.tick, 1000);
-  }
+  reset = () => {
+    this.setState({ timeLeft: LIMIT });
+  };
 
-  // componentDidUpdate で時間を確認して 0 になっていたら LIMIT に戻す
-  componentDidUpdate() {
+  tick = () => {
+    this.setState(prevState => ({ timeLeft: prevState.timeLeft - 1 }));
+  };
+
+  componentDidMount = () => {
+    this.timerId = setInterval(this.tick, 1000);
+  };
+
+  componentDidUpdate = () => {
     const { timeLeft } = this.state;
     if (timeLeft === 0) {
       this.reset();
     }
-  }
-
-  // componentWillUnmount で clearInterval
-  componentWillUnmount() {
-    clearInterval(this.timerId as NodeJS.Timer);
-  }
-
-  // reset で this.state.timeLeft に LIMIT を代入
-  reset = () => {
-    this.setState(() => ({ timeLeft: LIMIT }));
   };
 
-  // tick で this.state.timeLeft を減算（1秒経過）
-  tick = () => {
-    this.setState(prev => ({ timeLeft: prev.timeLeft - 1 }));
+  componentWillUnmount = () => {
+    clearInterval(this.timerId as NodeJS.Timer);
   };
 
   render() {
     const { timeLeft } = this.state;
 
     return (
-      <>
-        <TimerPage timeLeft={timeLeft} reset={this.reset} />
-      </>
+      <div className="container">
+        <header>
+          <h1>タイマー</h1>
+        </header>
+        <Card>
+          <Statistic className="number-board">
+            <Statistic.Label>time</Statistic.Label>
+            <Statistic.Value>{timeLeft}</Statistic.Value>
+          </Statistic>
+          <Card.Content>
+            <Button color="red" fluid onClick={this.reset}>
+              <Icon name="redo" />
+              Reset
+            </Button>
+          </Card.Content>
+        </Card>
+      </div>
     );
   }
 }
